% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tiles.R
\name{add_search_osm}
\alias{add_search_osm}
\title{Add search OSM to leaflet map}
\usage{
add_search_osm(.map, .position = "topleft")
}
\arguments{
\item{.map}{Leaflet map object.}

\item{.position}{Character value. One of "topleft", "topright", "bottomleft", or "bottomright".}
}
\description{
Add search OSM to leaflet map
}
