% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tiles.R
\name{add_tiles}
\alias{add_tiles}
\title{Add tiles to leaflet map}
\usage{
add_tiles(.map, .position = "bottomright")
}
\arguments{
\item{.map}{Leaflet map object.}

\item{.position}{Character value. One of "topleft", "topright", "bottomleft", or "bottomright".}
}
\description{
Add tiles to leaflet map
}
