% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_days.R
\name{aggregate_days}
\alias{aggregate_days}
\title{Aggregate dates in intervals}
\usage{
aggregate_days(.date, .interval = 5L)
}
\arguments{
\item{.date}{(vector of class Date) Dates to be aggregated.}

\item{.interval}{(integer value) The length of the interval (currently 5, 10, 15 supported).}
}
\value{
integer vector. For pentades a number between 1 and 72, e.g. "2016-02-16" will be 10.
}
\description{
For each month all days inside the interval will be aggregated, e.g. for an
interval length of five 1st January till 5th January will all belong to the first interval,
6th to 10th January will belong to interval 2 and so on.
Note that the last interval has a varying number of days,
e.g. 26th July to 31th July, but only 26th April to 30th April.
}
\examples{
aggregate_days(c("2017-02-16"), .interval = 10) # 5
aggregate_days(c("2016-01-31")) # 6
aggregate_days(c(NA, "2014-07-28")) # 42
}
