% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_date.R
\name{check_valid_date}
\alias{check_valid_date}
\title{Check date for validity}
\usage{
check_valid_date(.year = NULL, .month = NULL, .day = NULL)
}
\arguments{
\item{.year}{Integer value.}

\item{.month}{Integer value.}

\item{.day}{Integer value.}
}
\value{
Returns NULL if date is ok, else a string with the error message.
}
\description{
Check date for validity
}
\examples{
# errors
check_valid_date(.year = NA)

date.future <- Sys.Date() + 1L
check_valid_date(.year = lubridate::year(date.future),
                 .month = lubridate::month(date.future),
                 .day = lubridate::day(date.future))

check_valid_date(.year = lubridate::year(date.future) + 1L)
check_valid_date(.year = lubridate::year(date.future),
                 .month = lubridate::month(date.future) + 1L)

check_valid_date(.year = 2016, .month = NA, .day = 10)
check_valid_date(.year = 2018, .month = 13)
check_valid_date(.year = 2018, .month = 12, .day = 32)

# ok
check_valid_date(.year = 2019, .month = 1, .day = 10)
check_valid_date(.year = 2019, .month = 10, .day = NA)
check_valid_date(.year = 2019, .month = NA, .day = NA)
}
