% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_to_vector.R
\name{data.frame_to_vector}
\alias{data.frame_to_vector}
\alias{data.table_to_vector}
\alias{data_table_to_vector}
\title{Convert data.table or data.frame to vector}
\usage{
data.frame_to_vector(.Data)

data.table_to_vector(.Data)

data_table_to_vector(.Data)
}
\arguments{
\item{.Data}{A data.frame or data.table.}
}
\value{
An unnamed vector.
}
\description{
Convert data.table or data.frame to vector
}
\examples{
library(data.table)
data_table_to_vector(data.table(a = 1, b = 2))
}
