% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_query}
\alias{db_query}
\alias{db_execute}
\title{Shorthands for dbGetQuery and dbExecute}
\usage{
db_query(...)

db_execute(...)
}
\description{
These functions assume that a database connection object \code{pool} exists in the
global environment!
}
