% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_info.R
\name{get_date_info}
\alias{get_date_info}
\title{Get date information (year, month, mday, yday)}
\usage{
get_date_info(.date, .output = "list")
}
\arguments{
\item{.date}{(vector of dates)}
}
\value{
list with named entries with information extracted from the date, e.g.
year, month, day of month and year day. If a year is a leapyear all dates after
29th February will have a yday minus 1, so e.g. 1.March will have the same yday
independent if it is leapyear or not.
}
\description{
Get date information (year, month, mday, yday)
}
\examples{
get_date_info(c("2013-05-04", "2016-01-09"))
get_date_info(c(NA, "2016-03-09"))

# Behaviour for leap years
get_date_info("2016-03-01")
get_date_info("2017-03-01")
}
