% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_email.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Check if an email address is valid.}
\usage{
is_valid_email(.x)
}
\arguments{
\item{.x}{character}
}
\description{
Check if an email address is valid.
}
