% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_popup_sightings.R
\name{make_popup_sightings}
\alias{make_popup_sightings}
\title{Make popup of last sightings}
\usage{
make_popup_sightings(.Data, .tk25.name = NULL, .n = 10)
}
\arguments{
\item{.Data}{(data.table)}

\item{.tk25.name}{(Character vector)\cr Names of TK25 to show in popup.}

\item{.n}{(Integer value) \cr How many observations to includde in popup.}
}
\value{
A character value.
}
\description{
Make popup of last sightings
}
\examples{
library(data.table)
Data <- data.table(Beobachter = c("A", "A", "B", "A", "C"),
                   Jahr = c(2018, 2018, 2018, 2016, 2016))
make_popup_sightings(Data, .tk25.name = "8039 Starnberg")
}
