% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar_page_with_text.R
\name{navbar_page_with_text}
\alias{navbar_page_with_text}
\alias{navbarPageWithText}
\title{NavbarPage which can contain html text}
\usage{
navbar_page_with_text(..., text)

navbarPageWithText(..., text)
}
\arguments{
\item{...}{Arguments to navbarPage}

\item{text}{HTML text (or actionButton ...)}
}
\description{
NavbarPage which can contain html text
}
