% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenogram.R
\name{phenogram}
\alias{phenogram}
\title{Phenogram}
\usage{
phenogram(.x, .x.months = NULL, .method = "days", ...)
}
\arguments{
\item{.x}{integer vector of ydays, pentades or months}

\item{.x.months}{integer vector of the corresponding months}

\item{.method}{character: one of "days", "pentades" or "months"}

\item{...}{arguments passed to \link[grid:gpar]{grid::gpar}}
}
\description{
Phenogram
}
\examples{
set.seed(213)
x <- round(rnorm(n = 100, mean = 180, sd = 50))
phenogram(x)

x <- round(rnorm(n = 100, mean = 40, sd = 10))
phenogram(x, method = "pentades")

x <- round(rnorm(n = 100, mean = 5, sd = 1))
phenogram(x, method = "months")

# with the sum of each month under the plot
x <- round(rnorm(n = 1000, mean = 5, sd = 1))
phenogram(x, .x.months = sample(1:12, 1000, replace = TRUE), method = "months")

}
