% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_toast.R
\name{show_toast}
\alias{show_toast}
\title{Show toast}
\usage{
show_toast(.text, .which = "error", .timeout = 5000)
}
\arguments{
\item{.text}{Text to display.}

\item{.which}{One of "error", "success" or "info".}

\item{.timeout}{How long should the toast be displayed?}
}
\description{
Just a wrapper around shinytoastr::toastr functions
}
